+++ com.rush.logistic.client.company-product/src/main/java/com/rush/logistic/client/company_product/domain/product/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package com.rush.logistic.client.company_product.domain.repository;
+package com.rush.logistic.client.company_product.domain.product.repository;
 
-import com.rush.logistic.client.company_product.domain.entity.Product;
+import com.rush.logistic.client.company_product.domain.product.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
